refactor: split compaction module into focused files
Split the 929-line compaction.ts into a compaction/ subdirectory with
10 focused files, each handling a single concern: types, config,
binary stripping, token estimation, split point detection, sliding
window, tool output clearing, summarization, prompts, and the main
orchestrator. Pure refactor with zero functional changes.
