+++ apps/server/src/agent/ai-sdk-agent.ts
@@ -18,7 +18,7 @@ import { buildFilesystemToolSet } from '../tools/filesystem/build-toolset'
 import { buildMemoryToolSet } from '../tools/memory/build-toolset'
 import type { ToolRegistry } from '../tools/tool-registry'
 import { CHAT_MODE_ALLOWED_TOOLS } from './chat-mode'
-import { createCompactionPrepareStep } from './compaction'
+import { createCompactionPrepareStep } from './compaction/prepare-step'
 import { createContextOverflowMiddleware } from './context-overflow-middleware'
 import { buildMcpServerSpecs, createMcpClients } from './mcp-builder'
 import { buildSystemPrompt } from './prompt'
