+++ apps/server/src/agent/compaction/prompt.ts
